Create Create a K-means clustering algorithm to group customers of a retail store based on their purchase history.
This is a basic implementation of the K-means clustering algorithm in Python. Make sure to replace the example data with our actual customer purchase history data. Also, we may need to preprocess our data before passing it to the algorithm, like normalizing it or handling missing values. Additionally, we might want to add methods to evaluate the quality of the clustering, such as the silhouette score or elbow method.